+++ src/VectorTraits/Impl/AVector128/WVectorTraits128AdvSimdAbstract.Y.cs
@@ -86,6 +86,27 @@ namespace Zyl.VectorTraits.Impl.AVector128 {
             return Statics.YClamp(value, amin, amax);
         }
 
+
+        /// <inheritdoc cref="IWVectorTraits128.YRoundToZero_AcceleratedTypes"/>
+        public override TypeCodeFlags YRoundToZero_AcceleratedTypes {
+            get {
+                return Statics.YRoundToZero_AcceleratedTypes;
+            }
+        }
+
+        /// <inheritdoc cref="IWVectorTraits128.YRoundToZero(Vector128{float})"/>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public override Vector128<float> YRoundToZero(Vector128<float> value) {
+            return Statics.YRoundToZero(value);
+        }
+
+        /// <inheritdoc cref="IWVectorTraits128.YRoundToZero(Vector128{double})"/>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public override Vector128<double> YRoundToZero(Vector128<double> value) {
+            return Statics.YRoundToZero(value);
+        }
+
+
         /// <inheritdoc cref="IWVectorTraits128.YNarrowSaturate_AcceleratedTypes"/>
         public override TypeCodeFlags YNarrowSaturate_AcceleratedTypes {
             get {