+++ src/VectorTraits/Impl/AVector128/WVectorTraits128AdvSimdB64.Y.cs
@@ -45,6 +45,21 @@ namespace Zyl.VectorTraits.Impl.AVector128 {
             }
 
 
+            /// <inheritdoc cref="IWVectorTraits128.YRoundToZero_AcceleratedTypes"/>
+            public static TypeCodeFlags YRoundToZero_AcceleratedTypes {
+                get {
+                    TypeCodeFlags rt = TypeCodeFlags.Single | TypeCodeFlags.Double;
+                    return rt;
+                }
+            }
+
+            /// <inheritdoc cref="IWVectorTraits128.YRoundToZero(Vector128{double})"/>
+            [MethodImpl(MethodImplOptions.AggressiveInlining)]
+            public static Vector128<double> YRoundToZero(Vector128<double> value) {
+                return AdvSimd.Arm64.RoundToZero(value);
+            }
+
+
 #endif // NET5_0_OR_GREATER
         }
     }