+++ src/VectorTraits/Impl/AVector128/WVectorTraits128AdvSimdB64Abstract.Y.cs
@@ -42,6 +42,20 @@ namespace Zyl.VectorTraits.Impl.AVector128 {
         }
 
 
+        /// <inheritdoc cref="IWVectorTraits128.YRoundToZero_AcceleratedTypes"/>
+        public override TypeCodeFlags YRoundToZero_AcceleratedTypes {
+            get {
+                return Statics.YRoundToZero_AcceleratedTypes;
+            }
+        }
+
+        /// <inheritdoc cref="IWVectorTraits128.YRoundToZero(Vector128{double})"/>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public override Vector128<double> YRoundToZero(Vector128<double> value) {
+            return Statics.YRoundToZero(value);
+        }
+
+
         /// <inheritdoc cref="IWVectorTraits128.YShuffleG2_AcceleratedTypes"/>
         public override TypeCodeFlags YShuffleG2_AcceleratedTypes {
             get {