+++ src/VectorTraits/Impl/AVector128/WVectorTraits128Base.cs
@@ -365,19 +365,6 @@ namespace Zyl.VectorTraits.Impl.AVector128 {
 #endif // NET7_0_OR_GREATER
             }
 
-#if NET7_0_OR_GREATER
-            /// <inheritdoc cref="IWVectorTraits128.YRoundToZero(Vector128{double})"/>
-            [MethodImpl(MethodImplOptions.AggressiveInlining)]
-            public static Vector128<double> YRoundToZero(Vector128<double> value) {
-                Vector128<double> signMask = Vector128Constants.Double_SignMask;
-                Vector128<double> valueAbs = Vector128.AndNot(value, signMask);
-                Vector128<double> signData = Vector128.BitwiseAnd(value, signMask);
-                Vector128<double> rt = Floor(valueAbs); // Vector128.Floor need .NET 5+ .
-                rt = Vector128.BitwiseOr(rt, signData);
-                return rt;
-            }
-#endif // NET7_0_OR_GREATER
-
 
             /// <inheritdoc cref="IWVectorTraits128.ConvertToSingle_AcceleratedTypes"/>
             public static TypeCodeFlags ConvertToSingle_AcceleratedTypes {